src/test/java/org/hdn/mathutil/test/MathUtilityTest.java
package org.hdn.mathutil.test;

import org.hdn.mathutil.core.MathUtility;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.function.Executable;

import static org.junit.jupiter.api.Assertions.*;

class MathUtilityTest {

    @BeforeEach
    void setUp() {
    }

    @AfterEach
    void tearDown() {
    }

    @Test
    public void testFactorialGivenRightArgumentReturnsWell()
    {
        assertEquals(1, MathUtility.getFactorial(0));
        assertEquals(1, MathUtility.getFactorial(1));
        assertEquals(2, MathUtility.getFactorial(2));
        assertEquals(6, MathUtility.getFactorial(3));
        assertEquals(120, MathUtility.getFactorial(5));
    }

    @Test
    public void test()
    {
        assertThrows(IllegalArgumentException.class, () -> MathUtility.getFactorial(21));
    }

}